Remove support for the before_const_remove hook
This hook did not work in Rails 6 in `zeitwerk` mode. I do not like it very much
because a class or module should not be aware of the way the user loads it.
Also, class and module objects are not constants!

A proper place for a callback like this would be the reloading code itself.

The hook defined in AR (the only existing in the code base) looks broken to me
because it only modifies the registry for the current thread, what about the
other registries? I might look into this later.